Extract settings provider bridge
