+++ service-worker.js
@@ -162,6 +162,7 @@ const APP_SHELL = [
   '/js/chat-prompt-context.js',
   '/js/chat-summaries.js',
   '/js/settings.js',
+  '/js/settings-provider-bridge.js',
   '/js/settings-sync-panel.js',
   '/js/feedback.js',
   '/js/tour.js',
