+++ tests/test-correctness-phase2.js
@@ -95,6 +95,7 @@ const pwaAppShellAssets = [
   '/js/startup-maintenance.js',
   '/js/startup-ui.js',
   '/js/ai-verdict-engine.js',
+  '/js/settings-provider-bridge.js',
   '/js/import-loader.js',
   '/js/import-file-input.js',
   '/js/import-drop-zone.js',
